created the coordinator class
THe coordinator class to a. Receive requests from the user to start the computation (the
method signature for this will come from your API), and return a
suitable result status back to the user
b. Request that the data storage component read integers from the
location specified by the user (remember, the number of integers
involved is potentially very large!)
c. Pass the integers to the compute component
d. Request that the data storage component write the results to the
output