Data Storage Implementation (#25)
* DataStorage and ProcessAPI updates

DataStorage created for writing to file and reading from input, moved class and interface to it's own package. Updated ProcessAPIImpl with more verbose workflow and methods. Created package for processAPI test. Removed unused interfaces, we could reimplement something like DSCode again if we find a better use for it.

* Delimitter Support

There should be a way to get the delimitter from the input file, and the delimitter needs to be updated or passed each time we format results for output.

* checkstyle errors

brackets on ifs and elses, variable name in Data Storage.